Improve docker-compose-run Make helper
... and annotate it, because it's confusing.

Always set exit status, so it doesn't fall foul of `nounset` being set
for the executing shell.

Execute final `exit` in a subshell so it doesn't kill the current shell
(which would prevent subsequent commands). This means `errexit` will
behave as expected and can still be used to control fail-fastness.